requirements.txt
accelerate
dotenv
einops
# for gemini api
google-genai
gradio
huggingface_hub
imageio
ipdb
json_repair
matplotlib
nltk
numpy
omegaconf
openai
opencv-python>=4.7.0.72
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
qwen_vl_utils
requests
rich
sentencepiece
setuptools
sty
sympy
tabulate
tiktoken
timeout-decorator
timm
torch
torchvision
tqdm
transformers
typing_extensions
validators
xlsxwriter
torchvision
einops
